overllap fix needs image

+++ src/main/java/ap/ap_project/tempPlayer.java
@@ -87,7 +87,7 @@ public class tempPlayer {
         piece.setLayoutY(lY);
     }
 
-    public void move(int dc) {
+    public void move(double dc) {
 
         //System.out.print("in move");
